fix(DB/playercreateinfo_item): Amount of allowed item
* Increase the amount allowed from tinyint signed to smallint unsigned

+++ src/server/game/Globals/ObjectMgr.cpp
@@ -3333,7 +3333,7 @@ void ObjectMgr::LoadPlayerInfo()
                     continue;
                 }
 
-                int32 amount   = fields[3].GetInt8();
+                int32 amount   = fields[3].GetUInt16();
 
                 if (!amount)
                 {
